requirements.txt
beautifulsoup4==4.11.1
Django==3.2.16
django-appconf==1.0.5
django-admin-csvexport==2.2
django-blog-zinnia @ git+https://github.com/arrobalytics/django-blog-zinnia.git@993d0d9d81a9f4cfa3081185e952ef49bb0f4798
django-ckeditor==6.3.0
django-contrib-comments==2.2.0
django-imagekit==4.1.0
django-image-cropping==1.6.1
django-import-export==2.5.0
django-js-asset==1.2.2
django-tagging==0.5.0
django-tastypie==0.14.4
django-xmlrpc==0.1.8
easy_thumbnails==2.7.1
mots-vides==2015.5.11
mysqlclient
olefile==0.45.1
pilkit==2.0
Pillow==9.3.0
pyparsing==2.2.0
python-dateutil==2.6.1
python-mimeparse==1.6.0
pytz==2022.6
regex==2018.2.3
zinnia-wysiwyg-ckeditor==1.3
